+++ src/matrix/impl_permutation_mul.rs
