+++ src/matrix/mod.rs
@@ -24,7 +24,7 @@ mod deref;
 mod slice;
 mod base;
 mod permutation_matrix;
-mod permutation_matrix_mul;
+mod impl_permutation_mul;
 
 pub use self::base::{BaseMatrix, BaseMatrixMut};
 pub use self::permutation_matrix::{PermutationMatrix, Parity};
